Address PR Feedback #1
- Removed unused files (supporting-documents)
- Renamed BCAAPForm -> FormMetaData
- Removed unused dependencies (supporting-documents)
- Reformatted some files using Prettier
