--- server/src/BCAAPForm/bcaapform.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { BCAAPFormService } from './bcaapform.service';
-
-@Controller('forms')
-export class BCAAPFormController {
-  constructor(private bcaapFormService: BCAAPFormService) {}
-}
