--- server/src/BCAAPForm/bcaapform.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { BCAAPForm } from './bcaapform.entity';
-import { BCAAPFormService } from './bcaapform.service';
-import { BCAAPFormController } from './bcaapform.controller';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([BCAAPForm])],
-  exports: [],
-  controllers: [BCAAPFormController],
-  providers: [BCAAPFormService],
-})
-export class ApplicationModule {}
