+++ server/src/FormMetaData/formmetadata.controller.ts
@@ -1,0 +1,7 @@
+import { Controller } from '@nestjs/common';
+import { FormMetaDataService } from './formmetadata.service';
+
+@Controller('forms')
+export class FormMetaDataController {
+  constructor(private formMetaDataService: FormMetaDataService) {}
+}
