--- server/src/supporting-document/supporting-document.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { SupportingDocumentService } from './supporting-document.service';
-
-@Controller('supporting-documents')
-export class SupportingDocumentController {
-  constructor(private supportingDocumentService: SupportingDocumentService) {}
-}
