+++ server/src/user/user.entity.ts
@@ -1,5 +1,4 @@
-import { Application } from '@/application/application.entity';
-import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
+import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 import { CustomBaseEntity } from '../common/custom-base.entity';
 
 @Entity()
@@ -9,7 +8,4 @@ export class User extends CustomBaseEntity {
 
   @Column({ type: 'varchar', length: 100, nullable: false, unique: true })
   externalId: string;
-
-  @OneToMany(() => Application, (application) => application.user)
-  applications: Application[];
 }
